fedora: Update install scripts

+++ fedora/install/bitwarden-cli.sh
@@ -1,7 +1,17 @@
 #!/bin/sh
 
-curl -L -o bw-linux.zip "https://bitwarden.com"
+set -euo pipefail
+set -x
+
+if command -v bw-cli &> /dev/null; then
+	echo "bitwarden cli is already installed"
+	exit 0
+fi
+
+INSTALL_FILE=$HOME/.local/bin/bw-cli
+
+curl -L -o bw-linux.zip "https://bitwarden.com/download/?app=cli&platform=linux"
 unzip bw-linux.zip
 chmod u+x bw
-mv bw $HOME/.local/bin/bw-cli
+mv bw $INSTALL_FILE
 rm bw-linux.zip
